Filter out words that failed to match previously (fixes #8)
In the vast majority of browsers I tested, the overhead of keeping track of all matched words and filtering out those that failed to match is outweighed by the performance gain of less looping. http://jsperf.com/simplesearch-multi-word-match  The good news is that either way, we're cranking through several thousand matches per second.